requirements.txt
Django==1.11.7
djangorestframework==3.7.3
django-filter==1.1.0
djangorestframework-filters==0.10.2
pytz==2017.2
python-pushover==0.3